make tests better
